requirements.txt
langchain
langchain-google-genai
pymupdf
python-dotenv
trafilatura
